requirements.txt
streamlit
langchain
groq
langchain_groq
selenium
beautifulsoup4
lxml
html5lib
python-dotenv